requirements.txt
numpy==1.24.3
matplotlib==3.7.1
Pillow==9.5.0
scikit-image==0.20.0
opencv-python==4.7.0.72
streamlit==1.22.0
streamlit_jupyter==0.2.1
nbdev==2.3.12